switches the color get funcs to use uint8_t

+++ include/SDL_share.h
@@ -42,10 +42,10 @@ bool textureColor( SDL_Renderer *render, SDL_Rect loc, SDL_Texture *base, SDL_Te
 // query.c - getting data from existing objects
 // gets the offset'th pixel from surface->pixels
 void getSurfacePixel( SDL_Surface *src, unsigned x, unsigned y,
-		unsigned *r, unsigned *g, unsigned *b, unsigned *a );
+		uint8_t *r, uint8_t *g, uint8_t *b, uint8_t *a );
 // because it uses renderreadpixels, very slow
 void getRenderPixel( SDL_Renderer *rend, unsigned x, unsigned y,
-		unsigned *r, unsigned *g, unsigned *b, unsigned *a );
+		uint8_t *r, uint8_t *g, uint8_t *b, uint8_t *a );
 
 // text.c - the nightmare known as SDL text
 SDL_Surface *quickTextSurf( TTF_Font *font, SDL_Color fg, const char *text, int size, int width );
